+++ src/aui-scheduler/js/aui-scheduler-base-view.js
@@ -74,7 +74,7 @@ var SchedulerView = A.Component.create({
          * @type Number
          */
         height: {
-            value: 600
+            value: 650
         },
 
         /**