+++ src/aui-scheduler/js/aui-scheduler-view-table.js
@@ -116,6 +116,7 @@ var Lang = A.Lang,
     CSS_SVT_TABLE_DATA_EVENT_RIGHT = getCN(SCHEDULER_VIEW, TABLE, DATA, EVENT, RIGHT),
     CSS_SVT_TABLE_DATA_FIRST = getCN(SCHEDULER_VIEW, TABLE, DATA, FIRST),
     CSS_SVT_TABLE_GRID = getCN(SCHEDULER_VIEW, TABLE, GRID),
+    CSS_SVT_TABLE_GRID_CONTAINER = getCN(SCHEDULER_VIEW, TABLE, GRID, CONTAINER),
     CSS_SVT_TABLE_GRID_FIRST = getCN(SCHEDULER_VIEW, TABLE, GRID, FIRST),
 
     TPL_SVT_CONTAINER = '<div class="' + CSS_SVT_CONTAINER + '">' +
@@ -139,17 +140,18 @@ var Lang = A.Lang,
 
     TPL_SVT_MORE = '<a href="javascript:;" class="' + CSS_SVT_MORE + '">{labelPrefix} {count} {labelSuffix}</a>',
 
-    TPL_SVT_ROW = '<div class="' + CSS_SVT_ROW + '" style="top: {top}%; height: {height}%;"></div>',
+    TPL_SVT_ROW = '<div class="' + CSS_SVT_ROW + '"></div>',
 
     TPL_SVT_TABLE_DATA = '<table cellspacing="0" cellpadding="0" class="' + CSS_SVT_TABLE_DATA + '">' +
         '<tbody></tbody>' +
         '</table>',
 
-    TPL_SVT_TABLE_GRID = '<table cellspacing="0" cellpadding="0" class="' + CSS_SVT_TABLE_GRID + '">' +
+    TPL_SVT_TABLE_GRID = '<div class="' + CSS_SVT_TABLE_GRID_CONTAINER + '">' +
+        '<table cellspacing="0" cellpadding="0" class="' + CSS_SVT_TABLE_GRID + '">' +
         '<tbody>' +
         '<tr></tr>' +
         '</tbody>' +
-        '</table>',
+        '</table></div>',
 
     TPL_SVT_EV_ICON_LEFT = '<span class="' + [CSS_ICON, CSS_ICON_ARROWSTOP_LEFT].join(_SPACE) + '"></span>',
     TPL_SVT_EV_ICON_RIGHT = '<span class="' + [CSS_ICON, CSS_ICON_ARROWSTOP_RIGHT].join(_SPACE) + '"></span>',
@@ -450,6 +452,8 @@ var SchedulerTableView = A.Component.create({
             var rowRenderedColumns = 0;
             var rowNode = A.Node.create(TPL_SVT_TABLE_DATA_ROW);
 
+            var renderedEvents = false;
+
             instance.loopDates(rowStartDate, rowEndDate, function(celDate, index) {
                 var key = String(celDate.getTime());
 
@@ -473,7 +477,7 @@ var SchedulerTableView = A.Component.create({
                 var evtColNode = A.Node.create(TPL_SVT_TABLE_DATA_COL);
                 var evtNodeContainer = evtColNode.one(DIV);
 
-                if ((evtRenderedStack.length < events.length) && (rowDisplayIndex === (displayRows - 1))) {
+                if ((evtRenderedStack.length < events.length) && displayRows && (rowDisplayIndex === (displayRows - 1))) {
                     var strings = instance.get(STRINGS);
 
                     var showMoreEventsLink = A.Node.create(
@@ -489,6 +493,7 @@ var SchedulerTableView = A.Component.create({
                     showMoreEventsLink.setData(EVENTS, events);
 
                     evtNodeContainer.append(showMoreEventsLink);
+                    renderedEvents = true;
                 }
                 else if (evt) {
                     var evtSplitInfo = instance._getEvtSplitInfo(evt, celDate, rowStartDate, rowEndDate);
@@ -501,6 +506,7 @@ var SchedulerTableView = A.Component.create({
                     instance._syncEventNodeUI(evt, evtNodeContainer, celDate);
 
                     evtRenderedStack.push(evt);
+                    renderedEvents = true;
                 }
 
                 rowRenderedColumns++;
@@ -508,7 +514,7 @@ var SchedulerTableView = A.Component.create({
                 rowNode.append(evtColNode);
             });
 
-            return rowNode;
+            return renderedEvents ? rowNode : null;
         },
 
         /**
@@ -521,10 +527,11 @@ var SchedulerTableView = A.Component.create({
         buildEventsTable: function(rowStartDate, rowEndDate) {
             var instance = this,
                 displayRows = instance.get(DISPLAY_ROWS),
+                end = false,
                 intervalStartDate = DateMath.clearTime(instance._findCurrentIntervalStart()),
                 cacheKey = String(intervalStartDate.getTime()).concat(rowStartDate.getTime()).concat(rowEndDate.getTime()),
                 rowDataTableNode = instance.rowDataTableStack[cacheKey],
-                rowDisplayIndex;
+                rowDisplayIndex = 0;
 
             if (!rowDataTableNode) {
                 rowDataTableNode = A.Node.create(TPL_SVT_TABLE_DATA);
@@ -534,10 +541,20 @@ var SchedulerTableView = A.Component.create({
 
                 tableBody.append(titleRowNode);
 
-                for (rowDisplayIndex = 0; rowDisplayIndex < displayRows; rowDisplayIndex++) {
+                while (!end) {
                     var rowNode = instance.buildEventsRow(rowStartDate, rowEndDate, rowDisplayIndex);
 
-                    tableBody.append(rowNode);
+                    if (rowNode) {
+                        tableBody.append(rowNode);
+
+                        rowDisplayIndex++;
+                        if (displayRows && rowDisplayIndex >= displayRows) {
+                            end = true;
+                        }
+                    }
+                    else {
+                        end = true;
+                    }
                 }
 
                 instance.rowDataTableStack[cacheKey] = rowDataTableNode;
@@ -594,18 +611,9 @@ var SchedulerTableView = A.Component.create({
         buildGridRowNode: function(rowIndex) {
             var instance = this;
 
-            var displayRowsCount = instance._getDisplayRowsCount();
-            var rowRelativeHeight = 100 / displayRowsCount;
             var tableGridNode = instance._getTableGridNode(rowIndex);
 
-            var rowNode = A.Node.create(
-                Lang.sub(
-                    TPL_SVT_ROW, {
-                        height: rowRelativeHeight,
-                        top: rowRelativeHeight * rowIndex
-                    }
-                )
-            );
+            var rowNode = A.Node.create(TPL_SVT_ROW);
 
             rowNode.append(
                 tableGridNode.toggleClass(CSS_SVT_TABLE_GRID_FIRST, (rowIndex === 0))